C25OptimizationChallenge

Documentation for C25OptimizationChallenge.

C25OptimizationChallenge.C25OptimizationChallenge

C25OptimizationChallenge.CityProblemInst

C25OptimizationChallenge.CityProblem

C25OptimizationChallenge.breadth_first

C25OptimizationChallenge.breadth_first_parallel

C25OptimizationChallenge.generate_bound

C25OptimizationChallenge.generate_solution

C25OptimizationChallenge.naive_bound

C25OptimizationChallenge.partial_dfs

C25OptimizationChallenge.read_city_problem

C25OptimizationChallenge.repeat

C25OptimizationChallenge.select_node

C25OptimizationChallenge.small_bound

C25OptimizationChallenge.speed

C25OptimizationChallenge.total_distance

C25OptimizationChallenge.C25OptimizationChallenge — Module

C25OptimizationChallenge

This is a package for the Optimization Challenge for the C25 course at MIT.

C25OptimizationChallenge.CityProblemInst — Type

CityProblemInst

This stores an instance of the problem for internal use.

C25OptimizationChallenge.CityProblem — Method

CityProblem(city_string)

This function parses a city from the format of the HashCode2014 package and returns a CityProblemInst object.

C25OptimizationChallenge.breadth_first — Method

breadth_first(city; max_depth=5)

This function tries to find a good solution to the problem using a greedy approach with lookahead, using sequential computation of each of the car agendas.

C25OptimizationChallenge.breadth_first_parallel — Method

breadth_first_parallel(city; max_depth=5)

This function tries to find a good solution to the problem using a greedy approach with lookahead, using parallel computation of each of the car agendas.

C25OptimizationChallenge.generate_bound — Method

generate_bound(city)

This computes the best bound on the problem across all the different bounding methods.

C25OptimizationChallenge.generate_solution — Method

generate_solution(city, num_times=0, method=breadth_first, max_depth=5)

This function generates a solution to the problem using the current preferred method.

C25OptimizationChallenge.naive_bound — Method

naive_bound(city::CityProblemInst)

This generates a naive bound on the problem by just finding the total road length.

C25OptimizationChallenge.partial_dfs — Method

partial_dfs(graph, node, depth, visited_nodes, visited_edges, distances, times)

This function recursively calculates the score of a given node.

C25OptimizationChallenge.read_city_problem — Function

read_city_problem(city)

Generates a CityProblemInst for a given city.

C25OptimizationChallenge.repeat — Method

repeat(city, num_times, method, max_depth)

This function tries to find a good solution to the problem by repeatedly calling method and keeping the best solution found. This should result in improvements if method is randomized.

C25OptimizationChallenge.select_node — Method

select_node(current, candidates, graph, visited_nodes, visited_edges, distances, times)

This function selects the next node to traverse from a list of candidates by assigning each node a score and choosing the highest. The score is based on how good the best path starting with that node is.

C25OptimizationChallenge.small_bound — Method

small_bound(city::CityProblemInst)

This bounds the solution by repeatedly choosing the fastest street until the time is up.

C25OptimizationChallenge.speed — Method

speed(street)

Returns the speed of a street.

C25OptimizationChallenge.total_distance — Function

total_distance(solution, city)

This computes the total distance that a solution travels.