C25OptimizationChallenge

Documentation for C25OptimizationChallenge.

C25OptimizationChallenge.C25OptimizationChallenge

C25OptimizationChallenge.breadth_first

C25OptimizationChallenge.speed

C25OptimizationChallenge.C25OptimizationChallenge — Module

C25OptimizationChallenge

This is a package for the Optimization Challenge for the C25 course at MIT.

C25OptimizationChallenge.breadth_first — Method

breadth_first(city)

This function tries to find a good solution to the problem using something similar to a breadth-first search.

The algorithm used here is based on the random walk algorithm presented in class. There are two key differences:

There is a preference to take streets that have not been visited yet.

For a street that has not been visited yet, if there are multiple streets that have not been visited yet, the street with the highest speed is chosen.

The first of these prevents the algorithm from backtracking too much. The second of these is simply a heuristic that results in favorable behavior; it is generally better to travel faster.

C25OptimizationChallenge.speed — Method

speed(street)

Returns the speed of a street.